chore: split user interface into separate files
