+++ static/scripts/onboarding/github-oauth.ts
@@ -1,0 +1,11 @@
+import { User } from "./github-user";
+
+export interface OAuthToken {
+  provider_token: string;
+  access_token: string;
+  expires_in: number;
+  expires_at: number;
+  refresh_token: string;
+  token_type: string;
+  user: User;
+}
