+++ static/scripts/onboarding/github-user.ts
@@ -1,0 +1,51 @@
+export interface UserMetadata {
+  avatar_url: string;
+  email: string;
+  email_verified: boolean;
+  full_name: string;
+  iss: string;
+  name: string;
+  phone_verified: boolean;
+  preferred_username: string;
+  provider_id: string;
+  sub: string;
+  user_name: string;
+}
+
+export interface Identity {
+  id: string;
+  user_id: string;
+  identity_data: {
+    avatar_url: string;
+    email: string;
+    email_verified: boolean;
+    full_name: string;
+    iss: string;
+    name: string;
+    phone_verified: boolean;
+    preferred_username: string;
+    provider_id: string;
+    sub: string;
+    user_name: string;
+  };
+  provider: string;
+  last_sign_in_at: string;
+  created_at: string;
+  updated_at: string;
+}
+
+export interface User {
+  id: string;
+  aud: string;
+  role: string;
+  email: string;
+  email_confirmed_at: string;
+  phone: string;
+  confirmed_at: string;
+  last_sign_in_at: string;
+  app_metadata: { provider: string; providers: string[] };
+  user_metadata: UserMetadata;
+  identities: Array<Identity>;
+  created_at: string;
+  updated_at: string;
+}
